add update & delete clients

+++ routes/clients.js
@@ -17,6 +17,50 @@ const { body, validationResult } = require('express-validator');
  */
 
 /**
+ * Add a client
+ * @route POST /clients
+ * @group Clients
+ * @param {ClientDTO.model} client.body.required
+ * @consumes application/json
+ * @produces application/json
+ * @returns {ClientDTO.model} 200 - Client
+ * @returns {Error.model} 400 - You are not an admin
+ * @returns 401 - User not authentified
+ * @security JWT
+ */
+router.post('/', [
+	body('name').not().isEmpty(),
+	body('email').isEmail(),
+	body('phone').isMobilePhone(),
+	body('address').not().isEmpty(),
+],
+async (req, res, next) => {
+
+	const errors = validationResult(req);
+	if (!errors.isEmpty()) {
+		return res.status(400).json({
+			errors: errors.array() 
+		});
+	}
+
+	if(req.user.admin) {
+		let client = await sequelize.client.create({
+			name: req.body.name,
+			email: req.body.email,
+			phone: req.body.phone,
+			address: req.body.address,
+		});
+
+		res.json(client);
+	}
+	else {
+		res.status(400).json({
+			error: 'You are not an admin'
+		})
+	}
+});
+
+/**
  * @typedef UpdateClientDTO
  * @property {string} id - Id of the client
  * @property {string} name - Name of the client
@@ -52,6 +96,30 @@ async (req, res, next) => {
 			errors: errors.array() 
 		});
 	}
+
+	if(req.user.admin) {
+		await sequelize.client.update({
+			name: req.body.name,
+			email: req.body.email,
+			phone: req.body.phone,
+			address: req.body.address,
+		}, {
+			where: {
+				id: req.body.id
+			}
+		});
+
+		res.json(await sequelize.client.findOne({
+			where: {
+				id: req.body.id
+			}
+		}));
+	}
+	else {
+		res.status(400).json({
+			error: 'You are not an admin'
+		})
+	}
 });
 
 /**
@@ -67,7 +135,7 @@ async (req, res, next) => {
  * @consumes application/json
  * @produces application/json
  * @returns 200 - Project deleted
- * @returns {Error.model} 400 - This client doesn't exists
+ * @returns {Error.model} 400 - This client doesn't exists or you are not an admin
  * @returns 401 - User not authentified
  * @security JWT
  */
@@ -80,18 +148,25 @@ router.delete('/:client', async (req, res, next) => {
 		});
 	}
 
-	let client = await sequelize.client.findOne({
-		where: {
-			id: req.params.client
-		}
-	});
+	if(req.user.admin) {
+		let client = await sequelize.client.findOne({
+			where: {
+				id: req.params.client
+			}
+		});
 
-	if(client != null){
-		await client.destroy();
+		if(client != null){
+			await client.destroy();
+		}
+		else {
+			res.status(400).json({
+				error: 'This client doesn\'t exists'
+			})
+		}
 	}
 	else {
 		res.status(400).json({
-			error: 'This client doesn\'t exists'
+			error: 'You are not an admin'
 		})
 	}
 });
